// src/app/core/model/cargo.model.ts
export interface Cargo {
  idCargo: number;
  nombreCargo: string | null;
  codSucursal: number | null;
  idGerencia1: number | null;
  idGerencia2: number | null;
  idGerencia3: number | null;
  gerencia1: string | null;
  gerencia2: string | null;
  gerencia3: string | null;
  funciones: string | null;
  actions: string;
}